+++ Runtime/Tweens/Tween.cs
@@ -11,17 +11,17 @@ namespace Timespawn.EntityTween.Tweens
     {
         public static void Move(EntityManager entityManager, Entity entity, float3 start, float3 end, TweenParams tweenParams)
         {
-            Move(entityManager, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount);
+            Move(entityManager, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount, tweenParams.StartDelay);
         }
 
         public static void Move(EntityCommandBuffer commandBuffer, Entity entity, float3 start, float3 end, TweenParams tweenParams)
         {
-            Move(commandBuffer, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount);
+            Move(commandBuffer, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount, tweenParams.StartDelay);
         }
 
         public static void Move(EntityCommandBuffer.ParallelWriter parallelWriter, int sortKey, Entity entity, float3 start, float3 end, TweenParams tweenParams)
         {
-            Move(parallelWriter, sortKey, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount);
+            Move(parallelWriter, sortKey, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount, tweenParams.StartDelay);
         }
 
         public static void Move(
@@ -32,14 +32,15 @@ namespace Timespawn.EntityTween.Tweens
             float duration,
             EaseDesc easeDesc = default,
             bool isPingPong = false, 
-            int loopCount = 1)
+            int loopCount = 1,
+            float startDelay = 0.0f)
         {
             if (!CheckParams(easeDesc.Exponent, loopCount))
             {
                 return;
             }
 
-            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount);
+            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount, startDelay);
             entityManager.AddComponentData(entity, new TweenTranslationCommand(tweenParams, start, end));
         }
 
@@ -51,14 +52,15 @@ namespace Timespawn.EntityTween.Tweens
             float duration,
             EaseDesc easeDesc = default,
             bool isPingPong = false, 
-            int loopCount = 1)
+            int loopCount = 1,
+            float startDelay = 0.0f)
         {
             if (!CheckParams(easeDesc.Exponent, loopCount))
             {
                 return;
             }
 
-            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount);
+            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount, startDelay);
             commandBuffer.AddComponent(entity, new TweenTranslationCommand(tweenParams, start, end));
         }
 
@@ -71,30 +73,31 @@ namespace Timespawn.EntityTween.Tweens
             float duration,
             EaseDesc easeDesc = default,
             bool isPingPong = false, 
-            int loopCount = 1)
+            int loopCount = 1,
+            float startDelay = 0.0f)
         {
             if (!CheckParams(easeDesc.Exponent, loopCount))
             {
                 return;
             }
 
-            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount);
+            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount, startDelay);
             parallelWriter.AddComponent(sortKey, entity, new TweenTranslationCommand(tweenParams, start, end));
         }
 
         public static void Rotate(EntityManager entityManager, Entity entity, quaternion start, quaternion end, TweenParams tweenParams)
         {
-            Rotate(entityManager, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount);
+            Rotate(entityManager, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount, tweenParams.StartDelay);
         }
 
         public static void Rotate(EntityCommandBuffer commandBuffer, Entity entity, quaternion start, quaternion end, TweenParams tweenParams)
         {
-            Rotate(commandBuffer, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount);
+            Rotate(commandBuffer, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount, tweenParams.StartDelay);
         }
 
         public static void Rotate(EntityCommandBuffer.ParallelWriter parallelWriter, int sortKey, Entity entity, quaternion start, quaternion end, TweenParams tweenParams)
         {
-            Rotate(parallelWriter, sortKey, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount);
+            Rotate(parallelWriter, sortKey, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount, tweenParams.StartDelay);
         }
 
         public static void Rotate(
@@ -105,14 +108,15 @@ namespace Timespawn.EntityTween.Tweens
             float duration,
             EaseDesc easeDesc = default,
             bool isPingPong = false, 
-            int loopCount = 1)
+            int loopCount = 1,
+            float startDelay = 0.0f)
         {
             if (!CheckParams(easeDesc.Exponent, loopCount))
             {
                 return;
             }
 
-            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount);
+            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount, startDelay);
             entityManager.AddComponentData(entity, new TweenRotationCommand(tweenParams, start, end));
         }
 
@@ -124,14 +128,15 @@ namespace Timespawn.EntityTween.Tweens
             float duration,
             EaseDesc easeDesc = default,
             bool isPingPong = false, 
-            int loopCount = 1)
+            int loopCount = 1,
+            float startDelay = 0.0f)
         {
             if (!CheckParams(easeDesc.Exponent, loopCount))
             {
                 return;
             }
 
-            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount);
+            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount, startDelay);
             commandBuffer.AddComponent(entity, new TweenRotationCommand(tweenParams, start, end));
         }
 
@@ -144,30 +149,31 @@ namespace Timespawn.EntityTween.Tweens
             float duration,
             EaseDesc easeDesc = default,
             bool isPingPong = false, 
-            int loopCount = 1)
+            int loopCount = 1,
+            float startDelay = 0.0f)
         {
             if (!CheckParams(easeDesc.Exponent, loopCount))
             {
                 return;
             }
 
-            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount);
+            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount, startDelay);
             parallelWriter.AddComponent(sortKey, entity, new TweenRotationCommand(tweenParams, start, end));
         }
 
         public static void Scale(EntityManager entityManager, Entity entity, float3 start, float3 end, TweenParams tweenParams)
         {
-            Scale(entityManager, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount);
+            Scale(entityManager, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount, tweenParams.StartDelay);
         }
 
         public static void Scale(EntityCommandBuffer commandBuffer, Entity entity, float3 start, float3 end, TweenParams tweenParams)
         {
-            Scale(commandBuffer, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount);
+            Scale(commandBuffer, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount, tweenParams.StartDelay);
         }
 
         public static void Scale(EntityCommandBuffer.ParallelWriter parallelWriter, int sortKey, Entity entity, float3 start, float3 end, TweenParams tweenParams)
         {
-            Scale(parallelWriter, sortKey, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount);
+            Scale(parallelWriter, sortKey, entity, start, end, tweenParams.Duration, new EaseDesc(tweenParams.EaseType, tweenParams.EaseExponent), tweenParams.IsPingPong, tweenParams.LoopCount, tweenParams.StartDelay);
         }
 
         public static void Scale(
@@ -178,14 +184,15 @@ namespace Timespawn.EntityTween.Tweens
             float duration,
             EaseDesc easeDesc = default,
             bool isPingPong = false, 
-            int loopCount = 1)
+            int loopCount = 1,
+            float startDelay = 0.0f)
         {
             if (!CheckParams(easeDesc.Exponent, loopCount))
             {
                 return;
             }
 
-            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount);
+            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount, startDelay);
             entityManager.AddComponentData(entity, new TweenScaleCommand(tweenParams, start, end));
         }
 
@@ -197,14 +204,15 @@ namespace Timespawn.EntityTween.Tweens
             float duration,
             EaseDesc easeDesc = default,
             bool isPingPong = false, 
-            int loopCount = 1)
+            int loopCount = 1,
+            float startDelay = 0.0f)
         {
             if (!CheckParams(easeDesc.Exponent, loopCount))
             {
                 return;
             }
 
-            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount);
+            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount, startDelay);
             commandBuffer.AddComponent(entity, new TweenScaleCommand(tweenParams, start, end));
         }
 
@@ -217,14 +225,15 @@ namespace Timespawn.EntityTween.Tweens
             float duration,
             EaseDesc easeDesc = default,
             bool isPingPong = false, 
-            int loopCount = 1)
+            int loopCount = 1,
+            float startDelay = 0.0f)
         {
             if (!CheckParams(easeDesc.Exponent, loopCount))
             {
                 return;
             }
 
-            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount);
+            TweenParams tweenParams = new TweenParams(duration, easeDesc.Type, easeDesc.Exponent, isPingPong, loopCount, startDelay);
             parallelWriter.AddComponent(sortKey, entity, new TweenScaleCommand(tweenParams, start, end));
         }